Add trap handling
This is done via at_exit, because of some interactions between
interrupts, excon, and the heroku client's own exception handler.

Ancillary to this, use the implied-begin scope in Ruby methods to
scope the "ensure".  This results in a lot of indentation change but
little semantic change.  Consider running "git show -w".

Signed-off-by: Daniel Farina <daniel@heroku.com>